% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphPlotting.R
\name{mol_plot}
\alias{mol_plot}
\title{Plot molecular tbl_graph using ggraph}
\usage{
mol_plot(
  tbl_graph,
  my_colors = c(H = "grey80", C = "grey30", P = "blue", O = "red", N = "green", S =
    "orange", F = "cyan", Cl = "chartreuse", I = "darkmagenta")
)
}
\arguments{
\item{tbl_graph}{The input tbl_graph object representing the molecular graph.}

\item{my_colors}{A vector of colors to assign to different atom types. The
default color scheme includes predefined colors for common atom types H, C,
P, O, N, S, F, Cl, and I.}
}
\description{
This function plots a molecular \code{\link[tidygraph]{tbl_graph}} using the
\code{\link[ggraph]{ggraph}} package. The function generates a visualization
of the molecular graph, where nodes represent atoms and edges represent
bonds. The appearance of the plot can be customized using the provided color
scheme and van der Waals radii.
}
\examples{
\dontrun{
  ## NOT RUN
  molec <- ChemmineR::smiles2sdf("CCP(CO)CF(F)(F)")
  gr <- sdf2tbl_graph(molec)
  mol_plot(gr)
}

}
