% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphTransforming.R
\name{sdf2tbl_graph}
\alias{sdf2tbl_graph}
\title{Transform SDF format into a tbl_graph}
\usage{
sdf2tbl_graph(sdf)
}
\arguments{
\item{sdf}{The input SDF object in ChemmineR format.}
}
\value{
Returns a tbl_graph object representing the molecular graph.
}
\description{
Transforms \code{ChemmineR}'s \code{\link[ChemmineR]{SDFset}} format into a
\code{\link[tidygraph]{tbl_graph}} from \code{tidygraph}.
}
\details{
This function takes a \code{ChemmineR} \code{\link[ChemmineR]{SDFset}}
(Structure-Data File) format and converts it into a
\code{\link[tidygraph]{tbl_graph}} object from the \code{tidygraph} package.
The resulting \code{\link[tidygraph]{tbl_graph}} represents a molecular graph
with nodes as atoms and edges as bonds.
}
\examples{
\dontrun{
  ## NOT RUN
  molec <- ChemmineR::smiles2sdf("CCP(CO)CF(F)(F)")
  gr <- sdf2tbl_graph(molec)
}

}
