% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphProcessing.R
\name{subset_graph_at}
\alias{subset_graph_at}
\title{Subset molecular graph by focusing on the neighborhood around one atom}
\usage{
subset_graph_at(tbl_graph, focuspos = "P", loc_neigh = 3)
}
\arguments{
\item{tbl_graph}{The input tbl_graph object representing the molecular graph.}

\item{focuspos}{A character specifying the atom type to focus on.}

\item{loc_neigh}{A positive integer indicating how many bonds away an atom
can be from the central atom to be considered.}
}
\value{
Returns a tbl_graph object representing the subset of the molecular
  graph based on the neighborhood around the specified atom.
}
\description{
This function subsets the given \code{\link[tidygraph]{tbl_graph}} object by
focusing on the neighborhood around a specific atom. The resulting graph
includes only the atoms within a certain distance from the specified atom.
}
\examples{
\dontrun{
  ## NOT RUN
  molec <- ChemmineR::smiles2sdf("CCP(CO)CF(F)(F)")
  gr <- sdf2tbl_graph(molec)
  subs <- subset_graph_at(gr, focuspos = "P", loc_neigh = 1)
}

}
