% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphTransforming.R
\name{tbl_graph2mf}
\alias{tbl_graph2mf}
\title{Get molecular formula from molecular tbl_graph}
\usage{
tbl_graph2mf(tbl_graph)
}
\arguments{
\item{tbl_graph}{The input tbl_graph object representing the molecular graph.}
}
\value{
Returns the molecular formula as a character string.
}
\description{
This function calculates the molecular formula for a molecule represented as
a \code{\link[tidygraph]{tbl_graph}} from \code{tidygraph}.
}
\details{
This function calculates the molecular formula of a given
\code{\link[tidygraph]{tbl_graph}} object representing a molecular graph. The
resulting molecular formula provides information about the types and counts
of atoms present in the molecule. This function does not add implicit not
remove explicit hydrogen atoms.
}
\examples{
\dontrun{
  ## NOT RUN
  molec <- ChemmineR::smiles2sdf("CCP(CO)CF(F)(F)")
  gr <- sdf2tbl_graph(molec)
  mf <- tbl_graph2mf(gr)
}

}
