% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphTransforming.R
\name{tbl_graph2sdf}
\alias{tbl_graph2sdf}
\title{Transform a tbl_graph into SDF format}
\usage{
tbl_graph2sdf(tbl_graph)
}
\arguments{
\item{tbl_graph}{The input tbl_graph object representing the molecular graph.}
}
\value{
Returns an SDFset object in ChemmineR format.
}
\description{
Transform a \code{\link[tidygraph]{tbl_graph}} from \code{tidygraph} to
\code{ChemmineR} \code{\link[ChemmineR]{SDFset}} format
}
\details{
This function takes a \code{\link[tidygraph]{tbl_graph}} object from the
\code{tidygraph} package and converts it into \code{ChemmineR}'s
\code{\link[ChemmineR]{SDFset}} (Structure-Data File) format. The resulting
SDF object can be used for further analysis or storage.
}
\examples{
\dontrun{
  ## NOT RUN
  molec <- ChemmineR::smiles2sdf("CCP(CO)CF(F)(F)")
  gr <- sdf2tbl_graph(molec)
  molec2 <- tbl_graph2sdf(gr)
}

}
