% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphProcessing.R
\name{weight_bond_partners}
\alias{weight_bond_partners}
\title{Weighted bond partners around central node}
\usage{
weight_bond_partners(tbl_graph, center = 1, sphere = 1)
}
\arguments{
\item{tbl_graph}{The input tbl_graph object representing the molecular graph.}

\item{center}{An integer specifying the central node.}

\item{sphere}{An integer indicating how many nodes around the center should
be included.}
}
\value{
Returns a \code{data.frame} containing the symbols (atom types) and the
edge-weighted distances of the bond partners.
}
\description{
This function calculates the weighted bond partners around a central node in
a given \code{\link[tidygraph]{tbl_graph}} object. The weighted bond partners
are determined based on a specified sphere around the central node. The
resulting data frame includes the symbols (atom types) and the weighted
distances of the bond partners.
}
\examples{
\dontrun{
  ## NOT RUN
  molec <- ChemmineR::smiles2sdf("CCP(CO)CF(F)(F)")
  gr <- sdf2tbl_graph(molec)
  wbp <- weight_bond_partners(gr, center = 3, sphere = 1)
}

}
